Moves Grape::Endpoint extension into a module
Adding methods into the Grape::Endpoint class that
ActiveModel::Serializers are relying on from ActionController as well as
a few convenience methods for Route and Namespace options. I believe we
can expand on some of the empty methods here to allow for higher (api)
level option setting and defaults. This is a first step in that
direction.